requirements.txt
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
pandas==2.1.1
numpy==1.26.0
matplotlib==3.8.0
requests==2.31.0
pyyaml==6.0.1
click==8.1.7
python-dotenv==1.0.0
src==0.0.7
